fix(#3815): report Body, never MapKeys, where the rendering cannot be synthesized; one shared answer for all four sites; split the spelling half out (#1135)
